[refactor/#180] Carousel 로직 훅으로 분리
